Loan Controller in microservice

+++ loan-microservice/src/main/java/com/loanmicroservice/loanmicroservice/Repositories/LoanRepo.java
@@ -5,10 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.loanmicroservice.loanmicroservice.Models.Loan;
 
-
 @Repository
-public interface LoanRepo extends JpaRepository<Loan, Integer> {
-
-
+public interface LoanRepo extends JpaRepository<Loan, Long> {
 
 }
